add jabs-init command to package install, eqivalent to initialize_project.py script
