+++ src/jabs/cli/__init__.py
@@ -1,5 +1,5 @@
 from .progress_bar import cli_progress_bar
 
 __all__ = [
-    "progress_bar",
+    "cli_progress_bar",
 ]
